[audio] regen my audio apps for Harmony 3.3 MH3-9704

+++ apps/audio_signal_generator/firmware/src/config/sam_e70_xult_wm8904_ssc_wqvga/gfx/libaria/src/libaria_widget_bar_graph.c
@@ -132,7 +132,7 @@ static void getValueLabelMaxDrawRect(laBarGraphWidget* graph, GFX_Rect * rect)
             //Protect from overflow
             if (graph->minValue > -MAX_TICK_LABEL_VALUE)
             {
-                sprintf(strbuff, "%ld", graph->minValue);
+                sprintf(strbuff, "%d", (int) graph->minValue);
             }
             else
             {
@@ -149,7 +149,7 @@ static void getValueLabelMaxDrawRect(laBarGraphWidget* graph, GFX_Rect * rect)
         //Protect from overflow
         if (graph->maxValue < MAX_TICK_LABEL_VALUE) 
         {
-            sprintf(strbuff, "%ld", graph->maxValue);
+            sprintf(strbuff, "%d", (int) graph->maxValue);
         } else 
         {
             sprintf(strbuff, "---");
