+++ apps/audio_signal_generator/firmware/src/config/sam_e70_xult_wm8904_ssc_wqvga/gfx/libaria/src/libaria_widget_circular_gauge_skin_classic.c
@@ -140,7 +140,7 @@ void drawTickLabelsAtAngleWithValue(laCircularGaugeWidget * gauge,
         //Protect from overflow
         if (value < MAX_TICK_LABEL_VALUE)
         {
-            sprintf(strbuff, "%d", value);
+            sprintf(strbuff, "%d", (int) value);
         }
         else
         {
