+++ apps/audio_signal_generator/firmware/src/config/sam_e70_xult_wm8904_ssc_wqvga/gfx/libaria/src/libaria_widget_line_graph_skin_classic.c
@@ -131,7 +131,7 @@ static void getValueLabelMaxDrawRect(laLineGraphWidget* graph, GFX_Rect * rect)
             //Protect from overflow
             if (graph->minValue > -MAX_TICK_LABEL_VALUE)
             {
-                sprintf(strbuff, "%ld", graph->minValue);
+                sprintf(strbuff, "%d", (int) graph->minValue);
             }
             else
             {
@@ -148,7 +148,7 @@ static void getValueLabelMaxDrawRect(laLineGraphWidget* graph, GFX_Rect * rect)
         //Protect from overflow
         if (graph->maxValue < MAX_TICK_LABEL_VALUE) 
         {
-            sprintf(strbuff, "%ld", graph->maxValue);
+            sprintf(strbuff, "%d", (int) graph->maxValue);
         } else 
         {
             sprintf(strbuff, "---");
@@ -212,7 +212,7 @@ static void drawTickLabelWithValue(laLineGraphWidget* graph, GFX_Point tickPoint
     //Protect from overflow
     if (value < MAX_TICK_LABEL_VALUE)
     {
-        sprintf(strbuff, "%ld", value);
+        sprintf(strbuff, "%d", (int) value);
     }
     else
     {
