+++ apps/audio_signal_generator/firmware/src/config/sam_e70_xult_wm8904_ssc_wqvga/gfx/libaria/src/libaria_widget_pie_chart_skin_classic.c
@@ -214,7 +214,7 @@ static void drawPieChart(laPieChartWidget* chart)
                     //Protect from overflow
                     if (value < MAX_TICK_LABEL_VALUE)
                     {
-                        sprintf(strbuff, "%ld", value);
+                        sprintf(strbuff, "%d", (int) value);
                     }
                     else
                     {
